+++ python/dazl/util/events.py
@@ -5,29 +5,23 @@
 Simple callbacks manager.
 """
 
-import asyncio
-import inspect
 import reprlib
-from asyncio import ensure_future, Future, get_event_loop
-from functools import wraps
+from asyncio import ensure_future, get_event_loop, Future
 from typing import Dict, Optional, List, Callable, Any, Awaitable
 
-from dazl.util.asyncio_util import propagate
-from ..util.typing import safe_cast
-
 from .. import LOG
+from .asyncio_util import DeferredStartTask, to_coroutine
+from ..util.typing import safe_optional_cast
 
 
 class CallbackManager:
     """
     A manager for callbacks, either normal ones or coroutines.
     """
-    def __init__(self, normalize_name_fn=None):
+    def __init__(self):
         self._records = dict()  # type: Dict[str, CallbackRecord]
-        if normalize_name_fn is not None:
-            self._normalize_name = normalize_name_fn
 
-    def add_listener(self, name, callback: Callable[..., Any], filter_fn=None, optional=False) -> 'CallbackDisconnector':
+    def add_listener(self, name: str, callback: Callable[..., Any], filter_fn=None) -> 'CallbackDisconnector':
         """
         Add a listener to be invoked whenever ``for_listeners(name)`` is called.
 
@@ -39,13 +33,10 @@ class CallbackManager:
         :param filter_fn:
             An optional callback that returns `True` or `False` on whether the corresponding
             callback should be invoked. This cannot be a coroutine function.
-        :param optional:
-            Track whether this callback has actually been invoked or not.
         """
         if not callable(callback):
             raise ValueError('callback must be callable')
 
-        name = self._normalize_name(name)
         record = self._records.get(name)
         if record is None:
             record = CallbackRecord()
@@ -53,8 +44,7 @@ class CallbackManager:
         record.append(callback, filter_fn)
         return CallbackDisconnector(record, callback)
 
-    def remove_listener(self, name, callback) -> bool:
-        name = self._normalize_name(name)
+    def remove_listener(self, name: str, callback) -> bool:
         record = self._records.get(name)
         if record is not None:
             try:
@@ -65,7 +55,7 @@ class CallbackManager:
 
         return False
 
-    def for_listeners(self, name) -> Callable[..., List['CallbackRecord']]:
+    def for_listeners(self, name: str) -> Callable[..., List['CallbackRecord']]:
         """
         Returns a proxy callback that takes parameters that are passed to each callback as is.
         Any return values are gathered and returned as a list of Future in the same order as the
@@ -77,12 +67,8 @@ class CallbackManager:
             LOG.error('Tried to use %s as a key for looking up event handlers.', name)
             raise
 
-        name = self._normalize_name(name)
         cb = self._records.get(name)
         return cb if cb is not None else lambda *_, **__: []
-
-    def rewrite_keys(self, rewrite_fn):
-        self._records = {rewrite_fn(key): value for key, value in self._records.items()}
 
     def keys(self):
         return self._records
@@ -92,8 +78,6 @@ class CallbackManager:
 
     def __bool__(self):
         return bool(self._records)
-
-    _normalize_name = str
 
 
 class EventHandlerTracker:
@@ -108,8 +92,8 @@ class EventHandlerTracker:
 
     def __init__(self, callback: Callable[..., Any], filter_fn=None):
         self.original_callback = callback
-        self.callback = _normalize_callback_implementation(callback)
-        self.pendings = []  # type: List[_StartableCallable]
+        self.callback = to_coroutine(callback)
+        self.pendings = []  # type: List[DeferredStartTask]
 
     def invoke_eventually(self, *args, **kwargs) -> Awaitable[Any]:
         """
@@ -121,7 +105,8 @@ class EventHandlerTracker:
         :param kwargs: Keyword arguments to the callback.
         :return: A Future that is resolved when the underlying callback is invoked.
         """
-        sc = _StartableCallable(lambda: self.callback(*args, **kwargs), start=not self.pendings, name=repr(self.callback))
+        sc = DeferredStartTask(lambda: self.callback(*args, **kwargs),
+                               start=not self.pendings, name=repr(self.callback))
         self.pendings.append(sc)
         sc.add_done_callback(self._maybe_execute_next)
         return sc
@@ -139,98 +124,6 @@ class EventHandlerTracker:
 
         if self.pendings:
             self.pendings[0].start()
-
-
-class _StartableCallable:
-    """
-    A :class:`Future`-like object that invokes a callback to start a computation at a later point
-    in time.
-    """
-    _asyncio_future_blocking = False
-
-    def __init__(self, cb: Callable[[], Future], start=False, name=None):
-        if not callable(cb):
-            raise ValueError('cb must be callable')
-
-        if start:
-            self._future = cb()
-            self._callback = None
-        else:
-            self._future = get_event_loop().create_future()
-            self._callback = cb
-
-        # This is required to be compatible with the interface of a Future.
-        self._loop = self._future._loop
-        self._name = name or repr(cb)
-
-    def start(self) -> None:
-        """
-        Actually invoke the underlying function (but only if it hasn't been invoked already).
-        """
-        cb = self._callback
-        if cb is not None:
-            self._callback = None
-            propagate(from_=ensure_future(cb()), to=self._future)
-
-    def started(self) -> bool:
-        return self._callback is None
-
-    def cancelled(self) -> bool:
-        return self._future.cancelled()
-
-    def cancel(self) -> bool:
-        self._callback = None
-        return self._future.cancel()
-
-    def result(self) -> Any:
-        return self._future.result()
-
-    def done(self) -> bool:
-        return self._future.done()
-
-    def exception(self) -> Optional[BaseException]:
-        return self._future.exception()
-
-    def add_done_callback(self, callback: Callable[[Future], Any]) -> None:
-        self._future.add_done_callback(callback)
-
-    def __await__(self):
-        return self._future.__await__()
-
-    def __repr__(self):
-        state = self._future._state if self.started() else 'NOT_STARTED'
-        return f'StartableFuture({self._name!r}, state={state})'
-
-
-def _normalize_callback_implementation(callback: Callable[..., Any]) -> Callable[..., Future]:
-    """
-    Convert the allowable callback signatures to a function that
-    :return:
-    """
-    if not callable(callback):
-        raise ValueError('callback must be callable')
-
-    @wraps(callback)
-    def invoke_sync(*args, **kwargs) -> Future:
-        loop = asyncio.get_event_loop()
-
-        try:
-            result = callback(*args, **kwargs)
-            if isinstance(result, asyncio.Future):
-                return result
-            else:
-                future = loop.create_future()
-                future.set_result(result)
-                return future
-        except Exception as ex:
-            future = loop.create_future()
-            future.set_exception(ex)
-            return future
-
-    if inspect.iscoroutinefunction(callback):
-        return lambda *args, **kwargs: ensure_future(callback(*args, **kwargs))
-    else:
-        return invoke_sync
 
 
 class CallbackRecord:
@@ -273,11 +166,11 @@ class InflightTracker:
 
     @classmethod
     def deferred(cls, callback, args, kwargs):
-        future = asyncio.get_event_loop().create_future()
+        future = get_event_loop().create_future()
         return InflightTracker(future, callback, args, kwargs)
 
-    def __init__(self, future: Optional[Future], callback, args, kwargs):
-        self.future = safe_cast(Future, future) if future is not None else None
+    def __init__(self, future: 'Optional[Future]', callback, args, kwargs):
+        self.future = safe_optional_cast(Future, future)
         if callable(callback):
             self.callback = callback
         else:
@@ -287,7 +180,7 @@ class InflightTracker:
 
     def start(self):
         LOG.debug("We're starting a thing...")
-        return asyncio.ensure_future(self.callback(*self.args, **self.kwargs))
+        return ensure_future(self.callback(*self.args, **self.kwargs))
 
 
 class CallbackDisconnector:
